+++ changelog.md
@@ -1,3 +1,22 @@
+# Changelog
+##1.1.0
+### EDutils
+- xds importer
+- dials importer
+- EDimporter (base class for xds and dials)
+- pets:
+  - added negative sign already in `pets.uvw`
+  - added dyngo functionalities
+- rotate_exp
+  - added dataframe based `integrate` function
+- utilities :
+  - added rodrigues rotation matrix
+  - added convert `to_shelx`
+### blochwave
+- added dyngo support
+- changes to frame_generation
+- changes to load bloch
+
 ##1.0.7
 ### blochwave
 - felix wrapper from Bloch._solve_Felix